% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dist3d.R
\name{dist3d}
\alias{dist3d}
\title{3D histogram distribution of two variables}
\usage{
dist3d(x, y, break.func = c("Sturges", "scott", "FD"), breaks = NULL,
  col.bar = "white", colvar = NULL, col.point = plot3D::gg.col(100),
  clab = NULL, phi = 5, theta = 25, ...)
}
\arguments{
\item{x,y}{vectors with x and y values.}

\item{break.func}{a function to compute the vector of breakpoints (see ?hist).}

\item{breaks}{an integer specifying the number of breaks.}

\item{col.bar}{color for bars.}

\item{colvar}{the variable used for coloring points}

\item{col.point}{Color palette to be used for the colvar variable.
If col is NULL and colvar is specified, then a red-yellow-blue colorscheme (jet.col) will be used.
If col is NULL and colvar is not specified, then col will be grey.}

\item{clab}{Only if colkey = TRUE, the label to be written on top of the color key.}

\item{theta,phi}{The angles defining the viewing direction}

\item{...}{other arguments to be passed to plot3D::hist3D()}
}
\description{
Draw a 3D histogram distribution of points.
}
\examples{
data(quakes)
dist3d(quakes$long, quakes$lat, colvar=quakes$depth,
   breaks =30)
}
\seealso{
\code{\link[plot3D]{hist3D}}, \code{\link[graphics]{hist}}
}

