% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rgl_init.R
\name{rgl_init}
\alias{rgl_init}
\title{Initialize RGL device}
\usage{
rgl_init(new.device = FALSE, bg = "white", width = 640)
}
\arguments{
\item{new.device}{a logical value. If TRUE, creates a new RGL device}

\item{bg}{the background color of the device}

\item{width}{the width of the device}
}
\description{
Creates a new RGL device if requested or if there
 is no opened device
}
\examples{
\donttest{
# Load dataset
data(iris)
x <- iris$Sepal.Length
y <- iris$Petal.Length
z <- iris$Sepal.Width
# Scatter plot
require("rgl")
rgl_init()
rgl.spheres(x, y, z, r = 0.2, color = "grey")
}
}

