% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph3d.R
\name{graph3d}
\alias{graph3d}
\title{<Add Title>}
\usage{
graph3d(data, width = "100\%", height = "100\%", style = "surface",
  tooltip = TRUE, showLegend = TRUE, cameraPosition = list(horizontal
  = 1, vertical = 0.5, distance = 2.5), xMin = NULL, xMax = NULL,
  yMin = NULL, yMax = NULL, zMin = NULL, zMax = NULL,
  elementId = NULL)
}
\description{
<Add Description>
}
\examples{
dat <- data.frame(x = c(1,1,2,2), y = c(1,2,1,2), z = c(1,2,3,4))
graph3d(dat, style = "bar", zMin = 0)
}
