% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph3d.R
\name{graph3d}
\alias{graph3d}
\title{<Add Title>}
\usage{
graph3d(data = NULL, x = ~x, y = ~y, z = ~z, xlab = NULL,
  ylab = NULL, zlab = NULL, width = "100\%", height = "100\%",
  style = "surface", showPerspective = TRUE, showGrid = TRUE,
  showShadow = FALSE, keepAspectRatio = TRUE, verticalRatio = 0.5,
  tooltip = TRUE, showLegend = TRUE, cameraPosition = list(horizontal
  = 1, vertical = 0.5, distance = 2.8), xMin = NULL, xMax = NULL,
  yMin = NULL, yMax = NULL, zMin = NULL, zMax = NULL,
  elementId = NULL)
}
\description{
<Add Description>
}
\details{
See \url{https://visjs.github.io/vis-graph3d/docs/graph3d/index.html#Configuration_Options}.
}
\examples{
dat <- data.frame(x = c(1,1,2,2), y = c(1,2,1,2), z = c(1,2,3,4))
graph3d(dat, style = "bar", zMin = 0)
graph3d(dat, style = "bar", zMin = 0,
        tooltip = JS(c("function(xyz){",
                       "  var x = 'X: ' + xyz.x.toFixed(2);",
                       "  var y = 'Y: ' + xyz.y.toFixed(2);",
                       "  var z = 'Z: ' + xyz.z.toFixed(2);",
                       "  return  x + '<br/>' + y + '<br/>' + z;",
                       "}"))
)
dat <- expand.grid(
  x = seq(-4,4,length.out=100),
  y = seq(-4,4,length.out=100)
)
dat <- transform(dat, density = dnorm(x)*dnorm(y))
graph3d(dat, z = ~density, keepAspectRatio = FALSE, verticalRatio = 1)
}
